Initial creation and upload of master AHK lexicon
Initial commit of the master AHK lexicon.
This file will be used by the installer as well as internal functions of the program to create and alter the main ahk.api.
The current setup allows for information to be easily accessed and retrieved.
Commands are setup in a way that main commands, subcommands, and keywords can be removed with ease.
There are many duplicates in this file. That's done in case commands are removed. All duplicates are removed from the API file by the aforementioned script.